Move mapToConnectionReferences() to util/db
